refactor(reports): extract shared query-AST enums into reports_enums
